Don't die if writing data out fails, in some bizarre manner.
Yes, this is rather inelegant, and it's not yet
known how the other side might interpret
a partial failure. I suspect a better way to
go would be to prepend the data length to the
data block itself and thus only do one write
instead of two. This is, however, pure supposition.